[Lombok] Extract `LoggerGenerator` out of `kotlin` subpackage because now it operates on Java classes as well
