+++ plugins/lombok/lombok.k2/src/org/jetbrains/kotlin/lombok/k2/FirLombokCommonRegistrar.kt
@@ -7,11 +7,16 @@ package org.jetbrains.kotlin.lombok.k2
 
 import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrar
 import org.jetbrains.kotlin.lombok.k2.config.LombokService
+import org.jetbrains.kotlin.lombok.k2.generators.LoggerGenerator
+import org.jetbrains.kotlin.lombok.k2.generators.LombokConstructorsGenerator
 import java.io.File
 
 class FirLombokCommonRegistrar(private val lombokConfigFile: File?) : FirExtensionRegistrar() {
     override fun ExtensionRegistrarContext.configurePlugin() {
         +LombokService.getFactory(lombokConfigFile)
+        +::LoggerGenerator
+        +::LombokConstructorsGenerator
+
         registerDiagnosticContainers(LombokCliDiagnostics)
     }
 }
