+++ plugins/lombok/lombok.k2/src/org/jetbrains/kotlin/lombok/k2/FirLombokJavaRegistrar.kt
@@ -12,7 +12,6 @@ class FirLombokJavaRegistrar : FirExtensionRegistrar() {
     override fun ExtensionRegistrarContext.configurePlugin() {
         +::AccessorGenerator
         +::WithGenerator
-        +::LombokConstructorsGenerator
         +::BuilderGenerator
         +::SuperBuilderGenerator
         +::DeclarationWithValueAnnStatusTransformer
