+++ plugins/lombok/lombok.k2/src/org/jetbrains/kotlin/lombok/k2/checkers/FirLombokConflictingLogFieldChecker.kt
@@ -15,9 +15,8 @@ import org.jetbrains.kotlin.fir.declarations.utils.isCompanion
 import org.jetbrains.kotlin.fir.scopes.impl.declaredMemberScope
 import org.jetbrains.kotlin.lombok.k2.LombokFirDiagnostics
 import org.jetbrains.kotlin.lombok.k2.config.lombokService
-import org.jetbrains.kotlin.lombok.k2.generators.kotlin.isLogger
+import org.jetbrains.kotlin.lombok.k2.generators.isLogger
 import org.jetbrains.kotlin.lombok.k2.generators.kotlin.isRelevantForConflictsCheck
-import org.jetbrains.kotlin.lombok.utils.LombokNames
 import org.jetbrains.kotlin.name.Name
 
 /**
